requirements.txt
# Django Framework
Django==4.2.16

# AI Integration
google-generativeai>=0.3.0
anthropic>=0.7.0

# Tailscale integration and networking
requests>=2.31.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# Utilities
celery>=5.3.0
redis>=5.0.0

# API and web services
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Development tools
django-debug-toolbar>=4.2.0
